Base node uptime calulation on System#nanoTime()
Previously, the uptime of a Graylog node has been calculated using the difference
between the start time and the current time. If the system time changed during the
runtime of the node (e. g. by synchronisation using an NTP service), the uptime would
be incorrect and could even be negative.

The current implementation is using the System#nanoTime() ticker to calculate the uptime
since the JVM (or more precisely: since the plugin) has been started.