Add opt-out REST endpoints and service to disable sending stats
The UsageStatsOptOutResource has one endpoint to check the current
opt-out status and a second one to submit the opt-out. The opt-out
status is persisted in the cluster config so it is available on all
nodes.

The Node and Cluster periodicals now check on every run if sending
stats is enabled.

Fixes #1